requirements.txt
requests
pyperclip
sqlparse
sql-metadata
Pillow